Move nearest structure reference to utils
