+++ src/gemdat/utils.py
@@ -6,6 +6,41 @@ import warnings
 import numpy as np
 from pymatgen.core import Lattice, Structure
 from scipy import signal
+from scipy.spatial import cKDTree
+
+
+def nearest_structure_reference(
+        structure: Structure) -> tuple[cKDTree, list[int]]:
+    """Find distance and index of the nearest site of the structure for each
+    voxel using a KD-tree.
+
+    Parameters
+    ----------
+    structure : pymatgen.core.structure.Structure
+        Structure of the material to use as reference for nearest site
+
+    Returns
+    -------
+    kd_tree : scipy.spatial.cKDTree
+        KD-tree of the structure
+    periodic_ids : list[int]
+        List of ids corresponding to the closest site of the structure
+    """
+    # In order to accomodate the periodicity, include the images of the structure sites
+    periodic_structure = []
+    periodic_ids: list[int] = []
+    images = np.mgrid[-1:2, -1:2, -1:2].reshape(3, -1).T
+    for dx, dy, dz in images:
+        periodic_structure.extend(structure.frac_coords +
+                                  np.array([dx, dy, dz]))
+
+        # store the id of the site in the original structure
+        periodic_ids.extend(range(len(structure.cart_coords)))
+
+    # Create a KD-tree from the structure
+    kd_tree = cKDTree(periodic_structure)
+
+    return kd_tree, periodic_ids
 
 
 def ffill(arr: np.ndarray, fill_val: int = -1, axis=-1) -> np.ndarray:
